File system changed.
Previously all the category specific scripts were placed in the directory named
modules/, but there was a problem that we can not execute that scripts form our
main script, because there was an error that "No such file or direcotry". So in order
to remove that error, I palaced all the category specific scripts in the base directory
where main script is also present.